custom-ui-perl-conversion/backend/requirements.txt
Flask
Flask-Cors
ibm-watsonx-ai
ibm-watson-machine-learning
python-dotenv
Flask-HTTPAuth
flask-marshmallow
Flask-SQLAlchemy
langchain
numpy
pymilvus
Pympler
pyparsing
pypdf
PyPika
python-dateutil
python-docx
python-dotenv
python-json-logger
regex
requests
urllib3
Werkzeug
sentence-transformers
langchain-community


